@@ -11,44 +11,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.7",
-    "@angular/cdk": "^11.2.11",
-    "@angular/common": "~11.2.7",
-    "@angular/compiler": "~11.2.7",
-    "@angular/core": "~11.2.7",
-    "@angular/forms": "~11.2.7",
-    "@angular/material": "^11.2.11",
-    "@angular/platform-browser": "~11.2.7",
-    "@angular/platform-browser-dynamic": "~11.2.7",
-    "@angular/router": "~11.2.7",
+    "@angular/animations": "^15.2.10",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "^15.2.10",
+    "@angular/compiler": "^15.2.10",
+    "@angular/core": "^15.2.10",
+    "@angular/forms": "^15.2.10",
+    "@angular/material": "^13.3.9",
+    "@angular/platform-browser": "^15.2.10",
+    "@angular/platform-browser-dynamic": "^15.2.10",
+    "@angular/router": "^15.2.10",
     "@ngneat/until-destroy": "^8.0.4",
     "hammerjs": "^2.0.8",
-    "ngx-color-picker": "^11.0.0",
-    "ngx-event-modifiers": "^1.0.1",
-    "ngx-take-until-destroy": "^5.4.0",
+    "ngx-color-picker": "^13.0.0",
     "object-hash": "^2.1.1",
-    "rxjs": "~6.6.0",
+    "rxjs": "~6.6.7",
     "ts-md5": "^1.2.7",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.3"
+    "tslib": "^2.3.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.6",
-    "@angular/cli": "~11.2.6",
-    "@angular/compiler-cli": "~11.2.7",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.1.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "@angular-devkit/build-angular": "^15.2.11",
+    "@angular/cli": "^15.2.11",
+    "@angular/compiler-cli": "^15.2.10",
+    "@types/jasmine": "~3.10.3",
+    "@types/node": "^12.20.55",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~3.10.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.20",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.7.0",
     "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.1.5"
+    "ts-node": "~10.4.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.9.5"
   }
 }
