@@ -1,7 +1,6 @@
 import { Directive, ElementRef, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { untilDestroyed } from 'ngx-take-until-destroy';
-import { fromEvent, merge, of } from 'rxjs';
-import { delay, map, switchMap } from 'rxjs/operators';
+import { fromEvent, merge, of, Subject } from 'rxjs';
+import { delay, map, switchMap, takeUntil } from 'rxjs/operators';
 
 @Directive({
 	// tslint:disable-next-line:directive-selector
@@ -11,30 +10,34 @@ export class DelayedHoverDirective implements OnInit, OnDestroy {
 	@Input()
 	delay = "1500";
 
-	@Output('delayed-hover') hoverEvent = new EventEmitter();
+	@Output('delayed-hover') hoverEvent = new EventEmitter<void>();
+    private destroy$ = new Subject<void>();
 
-	constructor(private readonly element: ElementRef) {}
+	constructor(private readonly element: ElementRef<HTMLElement>) {}
 
 	ngOnInit() {
-		const hide$ = fromEvent(this.element.nativeElement, 'mouseleave').pipe(map(_ => false));
-		const show$ = fromEvent(this.element.nativeElement, 'mouseenter').pipe(map(_ => true));
+		const hide$ = fromEvent(this.element.nativeElement, 'mouseleave').pipe(map(() => false));
+		const show$ = fromEvent(this.element.nativeElement, 'mouseenter').pipe(map(() => true));
 
 		merge(hide$, show$)
 			.pipe(
-				untilDestroyed(this),
-				switchMap(show => {
-					if (!show) {
+				takeUntil(this.destroy$),
+				switchMap(isHovering => {
+					if (!isHovering) {
 						return of(false);
 					}
-					return of(true).pipe(delay(parseInt(this.delay)));
+					return of(true).pipe(delay(parseInt(this.delay,10)));
 				})
 			)
-			.subscribe(show => {
-				if (show) {
+			.subscribe(shouldEmit => {
+				if (shouldEmit) {
 					this.hoverEvent.emit();
 				}
 			});
 	}
 
-	ngOnDestroy() {}
+	ngOnDestroy() {
+        this.destroy$.next();
+        this.destroy$.complete();
+    }
 }
